requirements.txt
Django==1.9.1
django-sphinx==2.2.4
djangorestframework==3.3.2
django-compressor==2.0
rest-social-auth==0.4.1
python-social-auth==0.2.14
drf-nested-routers==0.11.1
pycrypto==2.6.1
tornado==4.3


#dj-database-url==0.3.0
#dj-static==0.0.6
#django-appconf==1.0.1
#oauthlib==1.0.3
#requests==2.9.1
#requests-oauthlib==0.6.0
